+++ docs/conf.py
@@ -25,9 +25,9 @@ project = 'moban'
 copyright = '2017-2020 Onni Software Ltd.'
 author = 'chfw'
 # The short X.Y version
-version = '0.8.0'
+version = '0.8.1'
 # The full version, including alpha/beta/rc tags
-release = '0.8.0'
+release = '0.8.1'
 
 # -- General configuration ---------------------------------------------------